+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - proxy: add `Proxy::try_send_to()` and `Proxy::request_to()`.
 - telemeter: support gzip.
 - telemeter/openmetrics: expose units of registered metrics.
+- telemeter: the `elfo_metrics_storage_shards` gauge metric.
 
 ### Changed
 - telemeter: a new sharded-by-threads storage, it increases perf and dramatically reduces contention.